Faculty Avaliability_02 & New DB
Added functionality that detects if a user already has an available timeslot entry for the semester and the webpage displays differently based on if there is an existing entry or not.
-- Added a new Db--
Since Heuroku doesn't support free hosting for Postgresql, I now added code that makes the database run in a docker container.  DB IS UP AND RUNNING AGAIN!